HospitalizacionBDG5/requirements.txt
asgiref==3.5.2
Django==3.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3